% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{bkmr_wrapper}
\alias{bkmr_wrapper}
\title{Fits a BKMR model with significance criteria: PIP and group-wise PIP}
\usage{
bkmr_wrapper(y, x, args = list())
}
\arguments{
\item{y}{A vector of outcome}

\item{x}{A matrix of predictors}

\item{args}{A list of arguments, see R function `bkmr::kmbayes()`}
}
\value{
A list of vectors whose values are between 0 and 1
\item{pip}{PIP for component-wise selection or conditional (with-in group)
PIP for hierarchical variable selection.}
\item{group_pip}{PIP for group-specific selection.}
\item{time}{elapsed time to fit the model.}
}
\description{
Fits a BKMR model with significance criteria: PIP and group-wise PIP
}
\section{Reference}{


  Bobb JF, Henn BC, Valeri L, Coull BA (2018). “Statistical software for
  analyzing the health effects of multiple concurrent exposures via Bayesian
  kernel machine regression.”Environ-mental
  Health,17(67).doi:10.1186/s12940-018-0413-y.
}

