% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{bma_wrapper}
\alias{bma_wrapper}
\title{Fits a linear model with Bayesian model selection with significance criteria:
PIP and posterior probability of nonzero coefficients being on one side of
zero.}
\usage{
bma_wrapper(y, x, args = list())
}
\arguments{
\item{y}{A vector of outcome}

\item{x}{A matrix of predictors}

\item{args}{A list of arguments see R function `BMA::bic.glm()`.}
}
\value{
A list of vectors whose values are between 0 and 1
\item{beta}{The smaller posterior probability of the coefficients being to
one side of zero: min(Pr(beta >0), Pr(beta<0)).}
\item{pip}{PIP of the
effect not being zero.}
\item{time}{elapsed time to fit the model.}
}
\description{
Fits a linear model with Bayesian model selection with significance criteria:
PIP and posterior probability of nonzero coefficients being on one side of
zero.
}
\section{Reference}{


  Raftery A, Hoeting J, Volinsky C, Painter I, Yeung KY (2021).BMA: Bayesian
  model averaging. R package version 3.18.15.
}

