% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvine.R
\name{cor2partial}
\alias{cor2partial}
\title{Convert a correlation matrix into a partial correlation matrix}
\usage{
cor2partial(r)
}
\arguments{
\item{r}{A correlation matrix}
}
\value{
A partial correlation matrix
}
\description{
Convert a correlation matrix into a partial correlation matrix
}
