% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvine.R
\name{cvine}
\alias{cvine}
\title{Citation:
Daniel Lewandowski, Dorota Kurowicka, Harry Joe,
Generating random correlation matrices based on vines and extended onion method,
Journal of Multivariate Analysis,
Volume 100, Issue 9,
2009,
Pages 1989-2001,
ISSN 0047-259X,
https://doi.org/10.1016/j.jmva.2009.04.008.
Reference code: https://stats.stackexchange.com/questions/124538/how-to-generate-a-large-full-rank-random-correlation-matrix-with-some-strong-cor
Nother useful reference: https://www.sciencedirect.com/science/article/pii/S0047259X05000886:
        generates a positive definite correlation matrix}
\usage{
cvine(d, alpha = 10, beta = 10, S = NULL, m = 100)
}
\description{
Citation:
Daniel Lewandowski, Dorota Kurowicka, Harry Joe,
Generating random correlation matrices based on vines and extended onion method,
Journal of Multivariate Analysis,
Volume 100, Issue 9,
2009,
Pages 1989-2001,
ISSN 0047-259X,
https://doi.org/10.1016/j.jmva.2009.04.008.
Reference code: https://stats.stackexchange.com/questions/124538/how-to-generate-a-large-full-rank-random-correlation-matrix-with-some-strong-cor
Nother useful reference: https://www.sciencedirect.com/science/article/pii/S0047259X05000886:
        generates a positive definite correlation matrix
}
