% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{EstPara}
\alias{EstPara}
\title{estimate the true proportion of ref.otu.tab}
\usage{
EstPara(ref.otu.tab)
}
\arguments{
\item{ref.otu.tab}{a matrix, the reference OTU count table (row - taxa, column - samples), serving as the template for synthetic sample generation.}
}
\value{
A list with the elements
\item{mu}{true proportion}
\item{ref.otu.tab}{absolute abundance table}
}
\description{
estimate the true proportion of ref.otu.tab
}
