% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature.dat}
\alias{feature.dat}
\title{demo feature table}
\format{
A data frame with rownames are 856 features(row) and 60 samples(column)
}
\source{
Where you got the data from or how it was generated.
}
\usage{
feature.dat
}
\description{
demo feature table
}
\details{
Any additional details or context about the dataset.
}
\keyword{datasets}
