% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fit}
\alias{fit}
\title{Fits the model to given data and gets a list of significance criterions}
\usage{
fit(mod, x, y)
}
\arguments{
\item{mod}{An InferenceModel object}

\item{x}{A matrix of predictors}

\item{y}{A vector of outcome}
}
\description{
Fits the model to given data and gets a list of significance criterions
}
