% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{genx}
\alias{genx}
\title{Generates a matrix of n observations of p predictors}
\usage{
genx(obj, n)
}
\arguments{
\item{obj}{A MixtureModel object.}

\item{n}{An integer, number of observations to generate.}
}
\value{
A (n x p) dataframe.
}
\description{
Generates a matrix of n observations of p predictors
}
