% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{geny}
\alias{geny}
\title{Generates a vector of outcomes}
\usage{
geny(obj, x)
}
\arguments{
\item{obj}{An OutcomeModel object}

\item{x}{An (n x p) matrix of predictors}
}
\value{
An n-vector of outcomes
}
\description{
Generates a vector of outcomes
}
