% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{glm_wrapper}
\alias{glm_wrapper}
\title{Fits a generalized linear model}
\usage{
glm_wrapper(y, x, args = list())
}
\arguments{
\item{y}{A vector of outcome}

\item{x}{A matrix of predictors}

\item{args}{A list of arguments see R `glm` function.}
}
\value{
A list
\item{pval}{The p-value of the linear main effect.}
\item{time}{elapsed time to fit the model.}
}
\description{
Fits a generalized linear model
}
