% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{InferenceModel}
\alias{InferenceModel}
\title{Statistical model that returns significance criterions}
\usage{
InferenceModel(model, name = NULL, ...)
}
\arguments{
\item{model}{A string of name of built-in statistical models or a function
implements a statistical model and returns a list of significance
criterions for each predictor. Built-in options include "bma", "bkmr", "ms",
"bws", "qgc", "fin", "glm".}

\item{name}{A string, name of the statistical model. Default to string input
of model.}

\item{...}{Additional keyword arguments for the statistical model}
}
\description{
This functions creates a wrapper function for a statistical model and its
applicable significance criterions. It finds relationships between a matrix
of predictors and a vector of outcomes using the statistical model, and
determines if the relationships are "significant" according to prespecified
criterions for that model.
}
