% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixSelect.R
\name{MixSelect}
\alias{MixSelect}
\title{Gibbs sampler for MixSelect regression, based on a decomposition of the regression
          surface in:
          - main effects
          - interactions with heredity constraint
          - nonlinear deviation distributed as projected GP
ArXiv: https://arxiv.org/abs/1911.01910
Email: ff31@duke.edu for questions}
\usage{
MixSelect(
  y,
  X,
  Z = NULL,
  nrun = 2000,
  burn = nrun/2,
  thin = 1,
  tau = 1,
  c = 0.1,
  exp_C = 2,
  heredity = "strong",
  lambda_prior = "cont_spike",
  na = F,
  k_na = NULL,
  list_na = NULL,
  sd_lambda = 0.1,
  lod = F,
  X_lod = NULL,
  lod_vec = NULL,
  verbose = T
)
}
\arguments{
\item{y}{response vector;}

\item{X}{predictor matrix (n x p) to be included in projected GP;}

\item{Z}{covariate adjustments matrix (n x q);}

\item{nrun}{number of iterations;}

\item{burn}{burn-in period;}

\item{thin}{thinning interval; lambda_prior = "cont_spike",}

\item{tau}{variance hyperparameter for continuous spike and slab on main effects and interactions}

\item{c}{shrinkage parameter for continuous spike and slab on main effects and interactions}

\item{exp_C}{exponent of exponential covariance, default is  (2.2) in paper}

\item{heredity}{"strong" or "weak"}

\item{lambda_prior}{"cont_spike" or "discrete_spike", the  latter is extremely slow and  not  suggested}

\item{na}{T if analysis with missing data}

\item{k_na}{number of latent factors for analysis with NAs, refer to Section 5.3 in paper
Defaule is 2*log(p) where p is number of variables included in the Factor Model}

\item{list_na}{vector of dimension p+q indicating which variables in X and Z need to be included in  Factor Model,
this is because maybe we don't want to use a factor model for all the covariates, or maybe there are some variables
that are not missing
If null, use all columns in X and Z}

\item{sd_lambda}{hyperparamenter for element of Lambda matrix}

\item{lod}{T if analysis with data below the Limit of Detection (LOD)}

\item{X_lod}{Matrix of dimension nxp, 1 if observation i,j has been observed  below LOD and 1  otherwise}

\item{lod_vec}{vector or dimension  px1 containing  the LOD for each  variable  in  X}

\item{verbose}{logical}
}
\description{
Gibbs sampler for MixSelect regression, based on a decomposition of the regression
          surface in:
          - main effects
          - interactions with heredity constraint
          - nonlinear deviation distributed as projected GP
ArXiv: https://arxiv.org/abs/1911.01910
Email: ff31@duke.edu for questions
}
