% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{mplot}
\alias{mplot}
\title{Visualize marginals and Gaussian copula correlations of simulated data}
\usage{
mplot(obj, split = TRUE)
}
\arguments{
\item{obj}{A MixtureModel object.}

\item{split}{A logical, whether to display numbers on half of the covariance
matrix.}
}
\value{
A 'ggplot2' graphics.
}
\description{
Visualize marginals and Gaussian copula correlations of simulated data
}
