% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpower-package.R
\docType{package}
\name{mpower-package}
\alias{mpower}
\alias{mpower-package}
\title{mpower: Power Analysis via Monte Carlo Simulation for Correlated Data}
\description{
A flexible framework for power analysis using Monte Carlo simulation for settings in which considerations of the correlations between predictors are important. Users can set up a data generative model that preserves dependence structures among predictors given existing data (continuous, binary, or ordinal). Users can also generate power curves to assess the trade-offs between sample size, effect size, and power of a design. This package includes several statistical models common in environmental mixtures studies. For more details and tutorials, see Nguyen et al. (2022) \href{https://arxiv.org/abs/2209.08036}{arXiv:2209.08036}.
}
\author{
\strong{Maintainer}: Phuc H. Nguyen \email{phuc.nguyen.rcran@gmail.com} (\href{https://orcid.org/0000-0002-6206-0194}{ORCID})

}
\keyword{internal}
