% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{OutcomeModel}
\alias{OutcomeModel}
\title{Outcome generator}
\usage{
OutcomeModel(f, family = "gaussian", sigma = 1, f_args = list())
}
\arguments{
\item{f}{A string that describes the relationships between the
predictors and outcome or a function that takes an input matrix and returns
a vector of outcome: \eqn{E(y|x) = g(f(x))} where g is a link function
that depends on the family argument.}

\item{family}{A string, "gaussian", "binomial", or "poisson" for continuous,
binary, or count outcomes.}

\item{sigma}{A number, Gaussian noise standard deviation if applicable.}

\item{f_args}{A named list of additional arguments to f}
}
\value{
An OutcomeModel object. Attributes: f: mean function, sigma: a number Gaussian
  observation noise, family: a string "gaussian" or "binomial".
}
\description{
This function creates a generative model of the outcome given a matrix of
predictors.
}
