% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvine.R
\name{partial}
\alias{partial}
\title{Partial correlations between elements in x and elements in y}
\usage{
partial(r, x, y)
}
\arguments{
\item{r}{A correlation matrix}

\item{x}{A vector of indices}

\item{y}{A vector of indices}
}
\value{
A partial correlation matrix
}
\description{
Partial correlations between elements in x and elements in y
}
