% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{plot.mpower_resampling_MixtureModel}
\alias{plot.mpower_resampling_MixtureModel}
\title{Plot the correlations and univariate marginals of the generative model}
\usage{
\method{plot}{mpower_resampling_MixtureModel}(obj, split = TRUE)
}
\arguments{
\item{obj}{A MixtureModel object}

\item{split}{A boolean for whether to display numeric values of correlations}
}
\description{
Plot the correlations and univariate marginals of the generative model
}
