% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{plot}
\alias{plot}
\title{Visualize marginals and Gaussian copula correlations of simulated data}
\usage{
plot(obj, split = TRUE)
}
\arguments{
\item{obj}{A MixtureModel object}
}
\value{
ggplot2 graphics
}
\description{
Visualize marginals and Gaussian copula correlations of simulated data
}
