% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{qgcomp_lin_wrapper}
\alias{qgcomp_lin_wrapper}
\title{Fits a linear Quantile G-Computation model with no interactions}
\usage{
qgcomp_lin_wrapper(y, x, args = list())
}
\arguments{
\item{y}{A vector of outcome}

\item{x}{A matrix of predictors}

\item{args}{A list of arguments see R function `qgcomp::qgcomp.noboot()`.}
}
\value{
A list
\item{pval}{The p-value of the combined effect, the
same for all predictors.}
\item{pos_weights}{Positive weights. See 'qgcomp' package.}
\item{neg_weights}{Negative weights. See 'qgcomp' package.}
\item{time}{elapsed time to fit the model.}
}
\description{
Fits a linear Quantile G-Computation model with no interactions
}
\section{Reference}{


  Keil AP, Buckley JP, O’Brien KM, Ferguson KK, Zhao S, White AJ (2020). “A
  Quantile-based g-computation  approach  to  addressing  the  effects of
  exposure  mixtures.”Environmental Health Perspectives, 128(4).
}

