% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{qmultinom}
\alias{qmultinom}
\title{Quantile function for the multinomial distribution, size = 1}
\usage{
qmultinom(p, probs)
}
\arguments{
\item{probs}{a vector of probabilities for each level}
}
\description{
Quantile function for the multinomial distribution, size = 1
}
