% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{scale_sigma}
\alias{scale_sigma}
\title{Rescale the noise variance of a Gaussian OutcomeModel to meet a given SNR}
\usage{
scale_sigma(snr, ymod, xmod, m = 5000)
}
\arguments{
\item{snr}{A SNR}

\item{ymod}{A OutcomeModel object to modify}

\item{xmod}{A MixtureModel object}

\item{m}{Number of MC samples to estimate the SNR of a proposed noise variance}
}
\value{
A new OutcomeModel object
}
\description{
Rescale the noise variance of a Gaussian OutcomeModel to meet a given SNR
}
