% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{set_value}
\alias{set_value}
\title{This function updates values in an OutcomeModel object}
\usage{
set_value(obj, name, value)
}
\arguments{
\item{obj}{An OutcomeModel object}

\item{name}{A string for name of the attribute to be changed}

\item{value}{An appropriate data type}
}
\description{
This function updates values in an OutcomeModel object
}
