% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_curve}
\alias{sim_curve}
\title{Power analysis for multiple settings using Monte Carlo simulation}
\usage{
sim_curve(
  xmod,
  ymod,
  imod,
  s = 100,
  n = 100,
  cores = 1,
  file = NULL,
  errorhandling = "stop",
  snr_iter = 10000
)
}
\arguments{
\item{xmod}{A MixtureModel object.}

\item{ymod}{One or a list of OutcomeModel object(s).}

\item{imod}{An InferenceModel object.}

\item{s}{An integer for number of Monte Carlo simulation.}

\item{n}{An interger or a vector of sample sizes.}

\item{cores}{An integer for number of processing cores. When cores > 1,
parallelism is automatically applied.}

\item{file}{A string, a file name with no extension to write samples to
periodically. By default write to an RDS file.}

\item{errorhandling}{A string "remove", "stop", or "pass". If an error occurs
in any iteration, remove that iteration (remove), return the error message
verbatim in the output (pass), or terminate the loop (stop). Default is
"remove". See R package `foreach` for more details.}
}
\value{
A SimCurve object. Attributes: s: a number, snr: a number or list of
  numbers, n: a number or list of numbers, xmod: a MixtureModel, ymod: one or
  a list of OutcomeModels, imod: an InferenceModel, sims: a list of simulation
  output matrices.
}
\description{
This function can be used to create power curves by calling
  sim_power() on combinations of its inputs.
}
