% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_power}
\alias{sim_power}
\title{Power analysis using Monte Carlo simulation}
\usage{
sim_power(
  xmod,
  ymod,
  imod,
  s = 100,
  n = 100,
  cores = 1,
  file = NULL,
  errorhandling = "stop",
  snr_iter = 10000
)
}
\arguments{
\item{xmod}{A MixtureModel object.}

\item{ymod}{An OutcomeModel object.}

\item{imod}{An InferenceModel object.}

\item{s}{An integer for number of Monte Carlo simulation.}

\item{n}{An interger for sample size in each simulation.}

\item{cores}{An integer for number of processing cores. When cores > 1,
parallelism is automatically applied.}

\item{file}{A string, a file name with no extension to write samples to
periodically. By default write to an RDS file.}

\item{errorhandling}{A string "remove", "stop", or "pass". If an error occurs
in any iteration, remove that iteration (remove), return the error message
verbatim in the output (pass), or terminate the loop (stop). Default is
"remove". See R package `foreach` for more details.}

\item{snr_iter}{An integer for number of Monte Carlo samples to estimate SNR}
}
\value{
A PowerSim object. Attributes: xmod, ymod, imod, s, n, simulation
  samples.
}
\description{
Power analysis using Monte Carlo simulation
}
