% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{summary_one_sim_one_thres}
\alias{summary_one_sim_one_thres}
\title{See summary()
Returns a data.frame. Each column is power of a predictor. A column for threshold.}
\usage{
summary_one_sim_one_thres(sim, crit, thres, digits, how, pivot = FALSE)
}
\description{
See summary()
Returns a data.frame. Each column is power of a predictor. A column for threshold.
}
