% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{estimate_snr}
\alias{estimate_snr}
\title{Monte Carlo approximation of the SNR}
\usage{
estimate_snr(ymod, xmod, m = 5000, R = 100)
}
\arguments{
\item{ymod}{A OutcomeModel object}

\item{xmod}{A MixtureModel object}

\item{m}{Number of MC samples}

\item{R}{Number of bootstrap replicates}
}
\value{
An estimate SNR and 95-percent CI.
}
\description{
Monte Carlo approximation of the SNR
}
